Store claimable account to db. Create if missing.
